// src/client/components/types.tsx
export type valuesState = {
  intro: "";
  month: "";
  day: "";
  year: "";
  eligible1: ""; //Original Medicare
  eligible2: ""; //Medicare benefits
  coverage1: ""; // tricare
  coverage2: ""; //Medicare red
  coverage3: ""; //Medicaid?
  // Who did you enroll through for your current coverage?
  coverage4: ""; // current coverage,
  // step 4
  firstName: "";
  lastName: "";
  lovedFirstName: "";
  lovedLastName: "";
  phone: "";
  email: "";
  address: "";
  zip: "";
};